Structure change - move tenant relevant objects to tenancy
