+++ Source/Configuration/TenantSourceIdentifierResolverType.cs
@@ -1,9 +1,11 @@
 // Copyright (c) Aksio Insurtech. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Aksio.IngressMiddleware;
+namespace Aksio.IngressMiddleware.Configuration;
 
-public class TenantResolutionRouteOptions
+public enum TenantSourceIdentifierResolverType
 {
-    public string RegularExpression { get; set; } = string.Empty;
+    None = 0,
+    Route = 1,
+    Claim = 2
 }
