+++ Source/Tenancy/ClaimsSourceIdentifierResolverOptions.cs
@@ -1,9 +1,8 @@
 // Copyright (c) Aksio Insurtech. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Aksio.IngressMiddleware;
+namespace Aksio.IngressMiddleware.Tenancy;
 
-public class RouteSourceIdentifierResolverOptions
+public class ClaimsSourceIdentifierResolverOptions
 {
-    public string RegularExpression { get; set; } = string.Empty;
 }
