+++ Source/Tenancy/ITenantSourceIdentifierResolver.cs
@@ -3,7 +3,7 @@
 
 using Aksio.IngressMiddleware.Configuration;
 
-namespace Aksio.IngressMiddleware;
+namespace Aksio.IngressMiddleware.Tenancy;
 
 
 public interface ITenantSourceIdentifierResolver
