+++ Source/Tenancy/TenantResolutionRouteOptions.cs
@@ -1,8 +1,9 @@
 // Copyright (c) Aksio Insurtech. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Aksio.IngressMiddleware.Security;
+namespace Aksio.IngressMiddleware.Tenancy;
 
-public class ClaimsSourceIdentifierResolverOptions
+public class TenantResolutionRouteOptions
 {
+    public string RegularExpression { get; set; } = string.Empty;
 }
